requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
gunicorn==20.1.0
Markdown==3.4.1
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.2
whitenoise==6.2.0
psycopg2==2.8.6